refactors in styles and types
