+++ apps/hubspot/src/components/ConnectedEntriesTable.styles.ts
@@ -1,0 +1,8 @@
+import { css } from 'emotion';
+
+export const styles = {
+  rightCell: css({
+    minWidth: 140,
+    textAlign: 'right',
+  }),
+};
