+++ apps/hubspot/src/locations/Page.tsx
@@ -1,48 +1,15 @@
 import React, { useEffect, useState } from 'react';
-import {
-  Box,
-  Table,
-  Text,
-  Spinner,
-  Flex,
-  Heading,
-  Button,
-  Badge,
-} from '@contentful/f36-components';
+import { Box, Text, Spinner, Flex, Heading } from '@contentful/f36-components';
 import { useSDK } from '@contentful/react-apps-toolkit';
-import { createClient } from 'contentful-management';
+import { createClient, EntryProps, KeyValueMap } from 'contentful-management';
 import ConfigEntryService from '../utils/ConfigEntryService';
 import { styles } from './Page.styles';
 import { ConnectedFields } from '../utils/utils';
-
-const getStatusBadge = (entry: any) => {
-  const isPublished = Boolean(entry.sys.publishedAt);
-  return (
-    <Badge variant={isPublished ? 'positive' : 'warning'}>
-      {isPublished ? 'Published' : 'Draft'}
-    </Badge>
-  );
-};
-
-const getLastUpdatedTime = (dateString: string | undefined) => {
-  if (!dateString) return '-';
-  const date = new Date(dateString);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-  if (diffHours < 24) {
-    return `${diffHours} hour${diffHours === 1 ? '' : 's'} ago`;
-  }
-  return date.toLocaleDateString('en-GB', {
-    day: '2-digit',
-    month: 'short',
-    year: 'numeric',
-  });
-};
+import ConnectedEntriesTable from '../components/ConnectedEntriesTable';
 
 const Page = () => {
   const sdk = useSDK();
-  const [entries, setEntries] = useState<any[]>([]);
+  const [entries, setEntries] = useState<EntryProps<KeyValueMap>[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const [connectedFields, setConnectedFields] = useState<ConnectedFields>({});
@@ -103,8 +70,8 @@ const Page = () => {
     fetchData();
   }, [sdk]);
 
-  const handleManageFields = (entry: any) => {
-    sdk.navigator.openEntry(entry.sys.id);
+  const handleManageFields = (entry: EntryProps<KeyValueMap>) => {
+    // todo : implement
   };
 
   return (
@@ -142,81 +109,6 @@ const Page = () => {
     </Flex>
   );
 };
-
-const ConnectedEntriesTable = ({
-  entries,
-  connectedFields,
-  displayFieldId,
-  locale,
-  onManageFields,
-}: {
-  entries: any[];
-  connectedFields: ConnectedFields;
-  displayFieldId: string;
-  locale: string;
-  onManageFields: (entry: any) => void;
-}) => (
-  <Box marginTop="spacingXl">
-    <Table>
-      <Table.Head>
-        <Table.Row>
-          <Table.Cell>
-            <Text fontWeight="fontWeightDemiBold">Entry name</Text>
-          </Table.Cell>
-          <Table.Cell>
-            <Text fontWeight="fontWeightDemiBold">Content type</Text>
-          </Table.Cell>
-          <Table.Cell>
-            <Text fontWeight="fontWeightDemiBold">Updated</Text>
-          </Table.Cell>
-          <Table.Cell>
-            <Text fontWeight="fontWeightDemiBold">Status</Text>
-          </Table.Cell>
-          <Table.Cell>
-            <Text fontWeight="fontWeightDemiBold">Connected fields</Text>
-          </Table.Cell>
-          <Table.Cell className={styles.rightCell}>
-            <Text fontColor="gray600" fontSize="fontSizeS">
-              Connected entries: {entries.length}/25
-            </Text>
-          </Table.Cell>
-        </Table.Row>
-      </Table.Head>
-      <Table.Body>
-        {entries.map((entry) => {
-          const name = entry.fields?.[displayFieldId]?.[locale] || 'Untitled';
-          const contentType = entry.sys.contentType.sys.name || entry.sys.contentType.sys.id;
-          const updated = getLastUpdatedTime(entry.sys.updatedAt);
-          const status = getStatusBadge(entry);
-          const connected = connectedFields[entry.sys.id] || [];
-          const connectedCount = connected.length;
-          return (
-            <Table.Row key={entry.sys.id}>
-              <Table.Cell>{name}</Table.Cell>
-              <Table.Cell>{contentType}</Table.Cell>
-              <Table.Cell>{updated}</Table.Cell>
-              <Table.Cell>{status}</Table.Cell>
-              <Table.Cell>
-                <Flex flexDirection="row" gap="spacingS" alignItems="center">
-                  <Text>{connectedCount}</Text>
-                </Flex>
-              </Table.Cell>
-              <Table.Cell className={styles.rightCell}>
-                <Button
-                  variant="secondary"
-                  size="small"
-                  onClick={() => onManageFields(entry)}
-                  testId={`manage-fields-${entry.sys.id}`}>
-                  Manage fields
-                </Button>
-              </Table.Cell>
-            </Table.Row>
-          );
-        })}
-      </Table.Body>
-    </Table>
-  </Box>
-);
 
 function EmptyState() {
   return (
